Daemon - teach audio-settings disable input to reduce inputs to 0
When the user sets the audio setting "enable_inputs" to false, the number of inputs is also set to 0. This appears to be required for scsynth to not work with audio inputs and skips the audio rate check which causes boot issues for people whose audio rates for input and output differ.