GUI - make phx webview "offtherecord"
This means no http cache or cookies are stored to local disk. Instead, the whole session is stored in memory and deleted when the app is closed.

Cookies and http cache may be needed in the future, but until there's a concrete case for storing the session to disk and the implications carefully considered, it makes sense to default to this mode.