Vendor - upgrade minitest
There is a function in later versions of minitest which I can use in
tests for upcoming code that exports the reference documentation to TOML files.
As such, upgrading to the latest version seems like a good idea.

There appear to be no severe/breaking changes mentioned in the minitest
changelogs, *and our existing test suite still passes with the upgrade*.

The only other changes in addition to the minitest upgrade are fixes for a few small
deprecation warnings.