API - remove Minecraft Pi API
For a while now we've supported a configureless API to Minecraft Pi Edition which is the free version of Minecraft that shipped with Raspberry Pi OS. We provided a set of functions for manipulating and controlling the Minecraft world designed to support teachers wanting to tap into live coding 3D gaming in addition to sound to engage their students in Computer Science.

Unfortunately Minecraft Pi Edition is no longer available on the latest Raspberry Pi OS, so our API is no longer useful as we specifically target the latest Raspberry Pi OS for our releases.

Older versions of Sonic Pi (with the built in API) are still available for older versions of Raspberry Pi OS that are still compatible with Minecraft Pi Edition.