GUI - disable public OSC UI if OSC server is disabled
Since the ability to turn on public OSC networking depends on the OSC server
itself being enabled, it does not make sense to allow the UI elements that
toggle public OSC networking to be enabled if the OSC server is currently
disabled.

Therefore, we now automatically turn 'off' and disable the public OSC networking
UI elements (preferences panel checkbox and main menu item) when the OSC server
itself is disabled.

There is now also a text label in the preferences panel that describes how to
turn the public OSC networking feature on, to make sure that it's clear why the
feature might appear disabled.