GUI - Qt6 - re-implement full screen mode
On Windows when in full-screen mode the menus and drop downs aren't visible due to limitations with OpenGL rendering. This could be fixed in Qt5 by setting QWindowsWindowFunctions::setHasBorderInFullScreen. However, this doesn't appear to be available in Qt 6.

Instead we simply drop the window chrome and set the size of the window to match the current monitor (whilst remembering the previous size). When we switch back out of full-screen mode we revert back to the previous size and position.

This needs to be tested on macOS and Linux