Daemon - evolve kill token into a more general comms token
Token is now a random 32 bit signed int (as per OSC specs). It's now used in all internal API OSC messages as a form of session ID/verification token. This makes it harder for arbitrary processes on the local machine to call the API.

The token is generated by the Daemon and passed to both the GUI and Tau. Token checks still need to be implemented for the SPider OSC API - however, for now the token is passed instead of the unused GUI id arg slot.