Daemon - make Tau restartable
Daemon now has the start of an API which includes receiving the keep_alive messages and now restart-tau messages which will kill Tau's BEAM process and restart it (inheriting the port numbers, comms token, etc).

Daemon no longer expects Tau to send keep-alive messages (this is now solely the job of the GUI process).

This is useful for development purposes where it's necessary to restart the BEAM process - this will now be possible without having to restart the whole application.