Daemon - increase resilience when restarting tau
Previously, if the user attempted to restart Tau multiple times in succession without waiting for the previous restart to complete, then it was possible to create a zombied erlang process. This is because the pid wasn't being correctly received and stored, so the ability to kill it to restart it was inhibited.

Still needs more testing and usage, but things are starting to feel a little more robust.