GUI - increase seize of separators and also highlight on mouse over
Whilst I prefer the thin look - it makes them less usable as the mouseover area is much smaller. This is even more problematic on touch screens.

It would be nice to have specific draggable "handle" areas but I can't quite seem to figure out how to do this. If anybody knows - I'd love to discuss!