Spider - ensure slept flag is only set to true when actually slept
When in a time_warp sleep is called to manipulate the logical clock and not actually perform a physical sleep.

Prior to this patch the following would incorrectly enter an infinite loop:

```
live_loop :foo do
  midi :e1
end
```

This is because `midi` uses `time_warp` internally which in turn uses `sleep`. Therefore `midi` incorrectly told Sonic Pi that the thread had slept which meant the checks performed within live_loop to ensure the thread had correctly slept or called sync allowed it to spin out of control.

With this patch the code above raises an exception after the first iteration as expected.