GUI - convert preferences to an overlay
Previously, the preferences was integrated to the right-hand-side column of windows but was hidden by default. This meant that on show, it would interfere with any manual width modifications made to the movable borders within the GUI. It would also render them non-modifiable whilst the preferences were open.

We now have the preferences effectively on a different layer that has no relationship with the layout of the right-hand-side GUI column. Instead, it slides in from the right above everything and slides away on hide. This was initially attempted to be implemented via layouts, but I couldn't find a way of getting the settings pane over to the right hand side using this approach. Instead, I used the move command to manually place the preferences based on its width and the width of the main GUI window. Care has also been taken to ensure this is dynamically updated if the window is resized.

This commit also includes some minor style tweaks for the preferences.