GUI - further tweaks to preference widget height
On macOS and Linux, the preferences were squashed vertically.

I attempted to fix this by specifying various size policies, but I suspect I have something conflicting getting in the way. For now, I just hacked the height to be at least the height of the settings widget plus some extra room and that at least alleviates the issue on macOS.

If someone knows a more appropriate and idiomatic way of achieving this, I'd love to know :-)