GUI - correctly restore visual position and internal geometry
This had been disabled - I think back when I was trying to detect a crasher - which is very likely to be the Lissajous issue recently fixed.

Now that's been fixed it makes sense to return back to the old functioning behaviour. If things start crashing again, it's possible this commit might be the culprit :-)