Spider - add new fns link and link_sync
These are highly experimental (and therefore are subject to change) functions that provide the means to synchronise a thread with the link metronome.

This doesn't just sync the BPM similar to use_bpm :link, it also syncs the phase of the beat - which means it will block the current beat similar to sleep or sync whilst it waits for the matching beat of the next bar (defaults to the first beat). The number of beats in a bar can also be changed (this defaults to 4).

There is also link_sync which first calls a sync waiting for the link play session to have started (or if it is currently playing skips the sync) and then calls link. This is useful if you want to "arm" a few live loops then initiate them via a Link call from one of the connected computers - perhaps even running Ableton Live.