Tau - attempt to improve smoothness of BPM transitions
Previously the sleep duration was determined at the start of the sleep period and fixed until the sleep period is over. This made sense up to the point where we enabled BPM changes to be stateful across threads (:link mode) and therefore have the possibility that the bpm changes mid sleep. This has the following ramifications:

* If the BPM increases whilst sleeping, it needs to wake up earlier than initially planned.
* If the BPM decreases whilst sleeping, it needs to go back to sleep for longer.

We now consider both of these cases by having a way of waking up sleeps in :link mode as well as sleeping again for 0 beats to ensure we don't need to sleep any more.

Note that this commit may introduce subtle and weird side effect bugs as it modifies the core sleep timing code.