Build - don't deploy assets on build
Need to find a way of working with all these build steps post v4. The issue currently is that tailwind (one of the css asset pipeline tools) isn't available for the default Raspberry Pi OS. We therefore need to find a way around this.

The current plan is to consider some sort of dev mode which does support dynamic css asset pipeline tools for development purposes and a prod mode which just uses pre-compiled assets.