GUI - fix recent regression which lost editor focus when using the escape key to dismiss the error pane
This was recently changed to avoid an odd bug which conflicted with the escape key shortcut for the new bpm scrubber. Just setting the focus (and not doing anything else the more "energetic" set focus function does) seems not to trigger the bug. The original set focus function can still be accessed via the shortcut and the menubar.