GUI - extract and display audio hardware information from scsynth logs
This is currently a bit brittle as scsynth renders different log information of different platforms. If the scsynth log format changes for any platform, this will need to be reflected in daemon.rb as appropriate. Ideally scsynth would make this information available in a standard format for parsing by external programs.