GUI - exit boot early and report when scsynth doesn't start correctly
Previously, we waited 30s to see if scsynth had started correctly - as in some cases (particularly on Windows) there were long launch times for new users.

However, now we short-circuit early if we detect the process has exited early and present a dialog to the user explaining as much as possible alongside the scsynth logfile. This is now in place of just giving the full aggregated log files and a generic error message after 30s.