Spider - use :link bpm mode by default
Given that the default BPM for link is 60, this should therefore not be a breaking change. Rather it enables users to modify the BPM of any of their default threads using the new metronome GUI interface. Previous behaviour can be explicitly obtained via `use_bpm 60` - this will detach the current thread from the link metronome and give it a static bpm of 60.