Spider - MIDI - fix midi_clock_beat duration arg
When moving the scheduling of the 24 distinct MIDI clock ticks to the BEAM, I forgot to account for the duration arg which lets you define how many Sonic Pi beats you want the MIDI beat ticks to be spread over. Also a bunch of scheduling logic remained which has now been removed.