Spider - experiment with different midi clock beat timing approach
MIDI clock beats are a bit stuttery on my RPi sending MIDI to an external synth via USB.

I'm not sure what the cause of this is - it could be issues with jitter in the USB comms or Erlang scheduler jitter.

This patch tries an alternative scheduling approach - instead of spawning 24 processes - one for each MIDI clock tick, we spawn just 1 process which iterates and sends each of the 24 ticks before terminating. It shouldn't make any difference but it's worth a try.