Daemon - don't disable hw inputs on macOS by default
Previously we were supplying the -I flag with 0 to scsynth on macOS. This was in an attempt to not bind to the audio inputs by default. However, with certain external sound cards, it appears that this also blocks output (perhaps they're duplex?).

We still set -i to 0 to specify no input channels, but no longer set the -I flag.

This only affects macOS.