GUI - improve scrubber drag behaviour
On some systems QCursor::setPos() doesn't appear to work which impacted the drag behaviour of the new scrubber widgets.

This patch assumes that setPos() is a null op and fixes this behaviour.

On systems where setPos() is working correctly, the mouse will still be reset to its original position when the dragging started. Otherwise, the final cursor position will move as dictated by the dragging.