GUI - further improvements to cross-platform scrubber behaviour
It appears that on some platforms QCursor::setPos(pos); is asynchronous and therefore calling QCursor::pos(); immediately after may not return the position just moved to.

We now detect whether setPos is functional on the current system at initialisation and then deviate our behaviour depending on this knowledge.

This stuff seems very fiddly to get to behave similarly across platforms!