Remove duplicate names in CONTRIBUTORS
I noticed that a couple of names had duplicate entries in the CONTRIBUTORS file, one with a blank author in the linked URL. I'm guessing this is an error and should be removed? Although I imagine you might have a script to update this so maybe there's a bug in there that's causing this?
There are several other names with a blank author in the linked URL, but since the name doesn't appear to be duplicated I've left them. However, it might be worth investigating how they happen and if they can be fixed (I thought maybe it could be if people have left GitHub since contributing?)