GUI - raise descriptive error message if home directory is not writable
Currently, Sonic Pi crashes if the directory determined to store .sonic-pi is not writable. This probably shouldn't be a problem and in the future it might make sense to teach Sonic Pi to be able to boot despite the home directory not being writable. However, for now, it makes much more sense to raise a clear descriptive error rather than fail silently. In the future, this boot error might be downgraded to a warning and rather than halting booting completely, continue with restrictions on things being saved etc.