API - separate Init and Boot methods
Previously the API Init method initialised all the paths, then booted the daemon and started the OSC server. However, the GUI now uses the API's path resolution to determine the config directory so it can read the audio config files and send the appropriate flag to the daemon to enable/disable audio inputs. Given the Init method both set paths and booted the daemon, this wasn't possible. Therefore, we Init just initialises the paths and other state, and Boot boots the servers. We can now Init, get the paths, figure out whether scsynth needs inputs, then pass that to the new Boot method.