Build - fix cmake build type arg passing for Windows
On macOS and Linux, it is just necessary to pass CMAKE_BUILD_TYPE during the Makefile generation proces (initiated by the config scripts in the app dir).

Unfortunately with Visual Studio on Windows this does not appear to work: https://stackoverflow.com/questions/24460486/cmake-build-type-is-not-being-used-in-cmakelists-txt

We therefore need to pass the build type to the --config flag for cmake when actually initiating a build. This was the previous behaviour which was changed in b4523374c9dbc3c98136da5d93d37b0317d6adb0 - this commit reverts this change. However, that change was introduced as part of an attempt to fix the build which was breaking under a Debug build type on Windows.

It turns out that the external projects (sp_midi and sp_link) were being hardcoded to a Release build which recently seems to have started causing issues on the GitHub CI server. In order to set the build type of these projects correctly we now also pass CMAKE_BUILD_TYPE during Makefile generation process (similar to macOS and Linux) - this isn't used to build the GUI, (see above) but it is used to obtain the correct flag value for --config for the cmake build sub scripts called by the main cmake script itself.

Phew! Hopefully this fixes the CI (for now...)