Tau - add experimental support for Hydra
Hydra is a live code-able video synth created by Olivia Jack. See: https://hydra.ojack.xyz/

This commit introduces experimental support for Hydra via the new hydra fn which takes a string representing a Hydra sketch:

```
hydra "osc(10,1,1).saturate( () => Math.sin(time) * 50 ).out(o0)"
```

This sketch is then rendered on the Tau output (and any connected browsers) with the same timing as if it was a call to synth or sample.  Subsequent calls to hydra will replace the sketch. You can therefore call hydra within a live loop to modulate the visuals in time.

Note that the QtWebEngine-powered Tau tab is now enabled (currently found in the help pane). This will make it hard to compile on platforms that don't have access to QtWebEngine.