Build - restrict windows deletion pattern
Unfortunately ? doesn't represent exactly one character in Windows batch scripts if it followed by another ? This means this script can delete files other than expected - as it did with me :(

Also, unfortunately I can't seem to replicate with other approaches. For example the following looked promising:

for /f "eol=: delims=" %%F in ('dir /b ^| findstr /r /e /c:".*-[a-f0-9]{32}.*"') do del "%%F"

Sadly findstr doesn't support {n} for repitition. No worries, we can simply inline it:

for /f "eol=: delims=" %%F in ('dir /b ^| findstr /r /e /c:".*-[a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9][a-f0-9].*"') do del "%%F"

Oh no - findstr runs out of memory.

So until a nice batch friendly solution can be found, we resort to manually specifying file names including -