Server - rework how the Pi modname is generated/used
The pi's modname is currently used for 2 things:

1. To determine the default sched ahead time (depending on whether it's pi 2, pi 3 or something later).
2. To generate the user-readable host name displayed in the bottom right of the GUI.

We now use a smarter way of determining the info for the Pi (code submitted by Robin Newman) rather than relying on an ever-growing list of codes. This is then used to generate the readable host name from which we can easily determine if it's a pi2, pi3 or something else.