Lang - teach live_loop to move like live_audio
This touches in_thread, with_fx, live_loop and the threading infrastructure - so is highly likely to have associated bugs.

I'm committing it with all the logging statements in place incase these are needed for debugging purposes in the future - I'll remove them in the next commit, but they'll be here in the history.

This was extremely tricky code to write which is why it's taken me so long to get round to implementing such an obvious fix to the language.